Commands and cli inheritance organized
