+++ decko/lib/decko/commands/rake_command.rb
@@ -1,0 +1,12 @@
+
+require "cardio/commands/command"
+warn "DECKOCMD #{__LINE__}:#{__FILE__}"
+
+module Decko
+  module Commands
+
+    class RakeCommand < Cardio::Commands::Command
+    end
+  end
+end
+
